Game Over Scene added
Added a game over screen that appears whenever the player runs out of HP.  It has 2 buttons, one to reload the level and one to go back to the main menu.  The code to bring up the screen is in the player script, so any needed changes once multiple levels are added should be minimal.